Pequenas melhorias
- Formulário em arquivo dedicado
- Correções de identação
- Correção de warning sobre variável $this->content->text
- Melhoria no rótulo do campo de data
