+++ lang/en/block_clearmessages.php
@@ -1,6 +1,6 @@
 <?php
 $string['pluginname']         = 'Clear Messages';
-$string['startdate']          = 'Date';
+$string['startdate']          = 'Clear messages before';
 $string['clearbutton']        = 'Clear';
 $string['no_messages_found']  = 'No messages found before that date.';
 $string['messages_cleared']   = 'Messages have been cleared.';
