+++ lang/pt_br/block_clearmessages.php
@@ -1,6 +1,6 @@
 <?php
 $string['pluginname']         = 'Limpar Mensagens';
-$string['startdate']          = 'Data';
+$string['startdate']          = 'Limpar mensagens antes de';
 $string['clearbutton']        = 'Limpar';
 $string['no_messages_found']  = 'Nenhuma mensagem encontrada antes dessa data.';
 $string['messages_cleared']   = 'As mensagens foram apagadas.';
